+++ nengo_dl/tests/test_graph_optimizer.py
@@ -1,5 +1,7 @@
 # pylint: disable=missing-docstring
 
+from distutils.version import LooseVersion
+
 from nengo.exceptions import BuildError
 from nengo.neurons import LIF, LIFRate, Izhikevich, AdaptiveLIF
 from nengo.synapses import Lowpass, Triangle, Alpha, LinearFilter
@@ -9,10 +11,12 @@ from nengo.builder.operator import (SimPyFunc, DotInc, Copy, Reset,
                                     ElementwiseInc)
 from nengo.builder.processes import SimProcess
 from nengo.builder.signal import Signal
+from nengo.version import version as nengo_version
 import numpy as np
 import pytest
 
 from nengo_dl import op_builders
+from nengo_dl.compat import SparseDotInc
 from nengo_dl.graph_optimizer import (
     mergeable, greedy_planner, tree_planner, transitive_planner, noop_planner,
     order_signals, noop_order_signals,
@@ -57,6 +61,20 @@ def test_mergeable():
     # first dimension mismatch
     assert mergeable(dummies.Op(sets=[dummies.Signal(shape=(3, 2))]),
                      [dummies.Op(sets=[dummies.Signal(shape=(4, 2))])])
+
+    # sparsity mismatch
+    assert not mergeable(dummies.Op(sets=[dummies.Signal(sparse=False)]),
+                         [dummies.Op(sets=[dummies.Signal(sparse=True)])])
+
+    # sparse matrices are allowed to have mismatched shapes
+    assert mergeable(
+        dummies.Op(sets=[dummies.Signal(shape=(1, 2), sparse=True)]),
+        [dummies.Op(sets=[dummies.Signal(shape=(1, 3), sparse=True)])])
+    assert mergeable(
+        dummies.Op(sets=[dummies.Signal(base_shape=(2, 2), shape=(4, 1),
+                                        sparse=True)]),
+        [dummies.Op(sets=[dummies.Signal(base_shape=(2, 2), shape=(1, 4),
+                                         sparse=True)])])
 
     # Copy (inc must match)
     assert mergeable(Copy(dummies.Signal(), dummies.Signal(), inc=True),
@@ -161,6 +179,16 @@ def test_mergeable():
     b = SimBCM(dummies.Signal((5,)), dummies.Signal(), dummies.Signal(),
                dummies.Signal(), dummies.Signal())
     assert not mergeable(a, [b])
+
+
+@pytest.mark.skipif(LooseVersion(nengo_version) <= "2.8.0",
+                    reason="Nengo Sparse transforms not implemented")
+def test_sparsedotinc_mergeable():
+    assert mergeable(
+        SparseDotInc(dummies.Signal(sparse=True), dummies.Signal(),
+                     dummies.Signal()),
+        [SparseDotInc(dummies.Signal(sparse=True), dummies.Signal(),
+                      dummies.Signal())])
 
 
 def test_planner_mergeable(planner):